backend/requirements.txt
fastapi==0.105.0
openai==1.6.1
pydantic==2.5.3
python-dotenv==1.0.0
Requests==2.31.0
SQLAlchemy==2.0.18
uvicorn==0.25.0
zhipuai==1.0.7
python-multipart==0.0.5
python-jose[cryptography]
passlib[bcrypt]
databases